+++ .gitignore
@@ -1,6 +1,6 @@
 coverage
 node_modules
-dist
+dist/
 .env
 build
 dist
