+++ bin/build.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+set -e
+
+for x in src/lambdas/*; do
+  (cd "$x" && webpack)
+done
