+++ fixtures/items.js
@@ -1,4 +1,4 @@
-const common = require('./common.js')
+const common = require('./common')
 
 module.exports = () => ({
   mappings: {
