--- lambdas/api/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "api",
-  "description": "The api lambda function for stac-server",
-  "main": "index.js",
-  "license": "MIT",
-  "ava": {
-    "files": [
-      "tests/*.js"
-    ],
-    "verbose": true,
-    "serial": true
-  },
-  "scripts": {
-    "build": "webpack",
-    "test": "ava"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "ava": "^3.5.0",
-    "aws-event-mocks": "0.0.0",
-    "copy-webpack-plugin": "^9.0.1",
-    "proxyquire": "^2.1.0",
-    "sinon": "^7.1.1",
-    "webpack": "^5.42.1",
-    "webpack-cli": "~3.3",
-    "zip-webpack-plugin": "^3.0.0"
-  }
-}
