--- lambdas/failed-ingest/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "failed-ingest",
-  "description": "Failed ingest lambda function for stac-server",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack",
-    "test": "ava"
-  },
-  "ava": {
-    "files": [
-      "tests/*.js"
-    ],
-    "verbose": true,
-    "serial": true
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "ava": "^3.5.0",
-    "webpack": "^5.42.1",
-    "webpack-cli": "~3.3",
-    "zip-webpack-plugin": "^3.0.0"
-  }
-}
