--- lambdas/ingest/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "ingest",
-  "description": "ingest lambda function for stac-server",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack",
-    "test": "ava"
-  },
-  "ava": {
-    "files": [
-      "tests/*.js"
-    ],
-    "verbose": true,
-    "serial": true
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "ava": "^3.5.0",
-    "aws-event-mocks": "0.0.0",
-    "proxyquire": "^2.1.0",
-    "sinon": "^7.1.1",
-    "webpack": "^5.42.1",
-    "webpack-cli": "~3.3",
-    "zip-webpack-plugin": "^3.0.0"
-  }
-}
