--- libs/index.js
@@ -1,10 +1,0 @@
-/* eslint-disable global-require */
-
-'use strict'
-
-module.exports = {
-  api: require('./api.js'),
-  es: require('./es.js'),
-  esStream: require('./esStream.js'),
-  ingest: require('./ingest.js')
-}
