+++ package.json
@@ -6,23 +6,19 @@
   "author": "Alireza Jazayeri, Matthew Hanson <matt.a.hanson@gmail.com>, Sean Harkins",
   "license": "MIT",
   "scripts": {
-    "postinstall": "cd lambdas/api; npm install; cd ../ingest; npm install; cd ../failed-ingest; npm install",
-    "build": "cd lambdas/api; npm run build; cd ../ingest; npm run build; cd ../failed-ingest; npm run build",
+    "build": "./bin/build.sh",
     "lint": "eslint . --ext .js",
     "lint-js-fix": "eslint . --ext .js --fix",
     "test": "npm run test:unit",
-    "test:unit": "ava tests/*.js",
-    "test:coverage": "c8 ava tests/*.js",
-    "test:integration": "./bin/integration-tests.sh",
+    "test:coverage": "c8 npm run test:unit",
+    "test:integration": "ava tests/integration/*.js",
+    "test:system": "./bin/system-tests.sh",
+    "test:unit": "ava tests/unit/*.js",
     "deploy": "sls deploy",
     "serve": "serverless offline start",
     "build-api-docs": "npm run widdershins --search false --language_tabs 'nodejs:NodeJS' 'python:Python' --summary ./packages/api-lib/api-spec.yaml -o ./docs/api.md & npm run shins --inline --logo ./docs/images/logo.png -o ./docs/index.html ./docs/api.md"
   },
   "ava": {
-    "files": [
-      "tests/*.js",
-      "tests/integration/test_*.js"
-    ],
     "verbose": true,
     "serial": true
   },
@@ -43,9 +39,11 @@
     "through2": "^3.0.1"
   },
   "devDependencies": {
-    "ava": "^3.5.0",
+    "ava": "^3.15.0",
+    "aws-event-mocks": "^0.0.0",
     "aws-sdk": "^2.382.0",
     "c8": "^7.11.0",
+    "copy-webpack-plugin": "^10.2.4",
     "eslint": "^7.2.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.9.0",
@@ -57,13 +55,16 @@
     "prettier": "^2.0.5",
     "prettier-eslint": "^11.0.0",
     "prettier-eslint-cli": "^5.0.0",
-    "proxyquire": "^2.1.0",
+    "proxyquire": "^2.1.3",
     "serverless": "^2.51.0",
     "serverless-offline": "^8.4.0",
     "serverless-pseudo-parameters": "^2.5.0",
     "shins": "^2.3.2-3",
-    "sinon": "^7.1.1",
-    "widdershins": "^4.0.1"
+    "sinon": "^7.5.0",
+    "webpack": "^5.69.1",
+    "webpack-cli": "^4.9.2",
+    "widdershins": "^4.0.1",
+    "zip-webpack-plugin": "^4.0.1"
   },
   "name": "stac-server"
 }
