+++ src/lib/index.js
@@ -1,0 +1,10 @@
+/* eslint-disable global-require */
+
+'use strict'
+
+module.exports = {
+  api: require('./api'),
+  es: require('./es'),
+  esStream: require('./esStream'),
+  ingest: require('./ingest')
+}
